mod 내스터디 count 추가
