+++ src/app/(member)/mypage/friends/page.tsx
@@ -1,0 +1,3 @@
+export default function friends() {
+  return <>내친구</>;
+}
