+++ src/app/(member)/mypage/page.tsx
@@ -18,26 +18,33 @@ export default async function page() {
   } = await supabase.auth.getUser();
   const userdata = await getUser({ id: user?.id });
   const joindata = await getJoinInfo({ id: user?.id });
-  console.log(joindata.bookmark.length);
+  console.log(joindata)
 
-  const studyCardItem = [
+  type StudyCardItem = {
+    label: string;
+    num: number;
+    icon: JSX.Element;
+    path: string;
+  };
+
+  const studyCardItem: StudyCardItem[] = [
     {
       label: '관심 스터디',
       num: joindata.bookmark.length,
       icon: <ScrapIcon stroke="#A180F4" className="fill-none" />,
-      path: '/study',
+      path: '/bookmark',
     },
     {
       label: '참여 스터디',
       num: joindata.studymember.length,
       icon: <BookmarkIcon className="fill-[#A180F4]" />,
-      path: '/bookmark',
+      path: '/study',
     },
     {
       label: '스터디 친구',
       num: joindata.friend.length,
       icon: <FriendsIcon className="fill-[#A180F4]" />,
-      path: '/friend',
+      path: '/friends',
     },
   ];
 
@@ -63,7 +70,7 @@ export default async function page() {
               ))}
             </div>
           </div>
-          <MypageSection />
+          <MypageSection {...joindata}/>
         </div>
       )}
       {!userdata && <p className="text-center">로그인 정보가 없습니다</p>}
